+++ kong/plugins/opentelemetry/handler.lua
@@ -5,7 +5,7 @@
 -- at https://konghq.com/enterprisesoftwarelicense/.
 -- [ END OF LICENSE 0867164ffc95e54f04670b5169c09574bdbd9bba ]
 
-local BatchQueue = require "kong.tools.batch_queue"
+local Queue = require "kong.tools.queue"
 local http = require "resty.http"
 local clone = require "table.clone"
 local otlp = require "kong.plugins.opentelemetry.otlp"
@@ -45,10 +45,6 @@ local DEFAULT_CONTENT_TYPE_HEADER = "application/x-protobuf"
 local DEFAULT_HEADERS = {
   [CONTENT_TYPE_HEADER_NAME] = DEFAULT_CONTENT_TYPE_HEADER
 }
-
--- worker-level spans queue
-local QUEUES = {} -- one queue per unique plugin config
-
 
 local function get_headers(conf_headers)
   if not conf_headers or conf_headers == null then
@@ -101,24 +97,6 @@ local function http_export(conf, spans)
   end
 
   return ok, err
-end
-
-
-local function process_span(span, queue)
-  if span.should_sample == false or kong.ctx.plugin.should_sample == false then
-    -- ignore
-    return
-  end
-
-  -- overwrite
-  local trace_id = kong.ctx.plugin.trace_id
-  if trace_id then
-    span.trace_id = trace_id
-  end
-
-  local pb_span = encode_span(span)
-
-  queue:add(pb_span)
 end
 
 
@@ -175,28 +153,28 @@ end
 function OpenTelemetryHandler:log(conf)
   ngx_log(ngx_DEBUG, _log_prefix, "total spans in current request: ", ngx.ctx.KONG_SPANS and #ngx.ctx.KONG_SPANS)
 
-  local queue_id = kong.plugin.get_id()
-  local q = QUEUES[queue_id]
-  if not q then
-    local process = function(entries)
-      return http_export(conf, entries)
-    end
-
-    local opts = {
-      batch_max_size = conf.batch_span_count,
-      process_delay  = conf.batch_flush_delay,
-    }
-
-    local err
-    q, err = BatchQueue.new("opentelemetry", process, opts)
-    if not q then
-      kong.log.err("could not create queue: ", err)
+  kong.tracing.process_span(function (span)
+    if span.should_sample == false or kong.ctx.plugin.should_sample == false then
+      -- ignore
       return
     end
-    QUEUES[queue_id] = q
-  end
 
-  kong.tracing.process_span(process_span, q)
+    -- overwrite
+    local trace_id = kong.ctx.plugin.trace_id
+    if trace_id then
+      span.trace_id = trace_id
+    end
+
+    local ok, err = Queue.enqueue(
+      Queue.get_params(conf),
+      http_export,
+      conf,
+      encode_span(span)
+    )
+    if not ok then
+      kong.log.err("Failed to enqueue span to log server: ", err)
+    end
+  end)
 end