+++ plugins-ee/statsd-advanced/kong/plugins/statsd-advanced/schema.lua
@@ -218,6 +218,7 @@ return {
           { consumer_identifier_default = { type = "string", required = true, default = "custom_id", one_of = CONSUMER_IDENTIFIERS }, },
           { service_identifier_default = { type = "string", required = true, default = "service_name_or_host", one_of = SERVICE_IDENTIFIERS }, },
           { workspace_identifier_default = { type = "string", required = true, default = "workspace_id", one_of = WORKSPACE_IDENTIFIERS }, },
+          { queue = typedefs.queue },
         },
       },
     },