+++ spec/01-unit/01-db/01-schema/11-declarative_config/02-process_auto_fields_spec.lua
@@ -182,12 +182,17 @@ describe("declarative config: process_auto_fields", function()
               config = {
                 http_endpoint = "https://example.com",
                 content_type = "application/json",
-                flush_timeout = 2,
                 keepalive = 60000,
                 method = "POST",
-                queue_size = 1,
-                retry_count = 10,
                 timeout = 10000,
+                queue = {
+                  initial_retry_delay = 0.01,
+                  max_batch_size = 1,
+                  max_entries = 10000,
+                  max_coalescing_delay = 1,
+                  max_retry_delay = 60,
+                  max_retry_time = 60,
+                },
               }
             },
           }
@@ -235,12 +240,17 @@ describe("declarative config: process_auto_fields", function()
               config = {
                 http_endpoint = "https://example.com",
                 content_type = "application/json",
-                flush_timeout = 2,
                 keepalive = 60000,
                 method = "POST",
-                queue_size = 1,
-                retry_count = 10,
                 timeout = 10000,
+                queue = {
+                  initial_retry_delay = 0.01,
+                  max_batch_size = 1,
+                  max_entries = 10000,
+                  max_coalescing_delay = 1,
+                  max_retry_delay = 60,
+                  max_retry_time = 60,
+                },
               }
             },
           }
@@ -349,12 +359,17 @@ describe("declarative config: process_auto_fields", function()
                     config = {
                       http_endpoint = "https://example.com",
                       content_type = "application/json",
-                      flush_timeout = 2,
                       keepalive = 60000,
                       method = "POST",
-                      queue_size = 1,
-                      retry_count = 10,
                       timeout = 10000,
+                      queue = {
+                        initial_retry_delay = 0.01,
+                        max_batch_size = 1,
+                        max_entries = 10000,
+                        max_coalescing_delay = 1,
+                        max_retry_delay = 60,
+                        max_retry_time = 60,
+                      },
                     }
                   },
                 }
@@ -668,12 +683,17 @@ describe("declarative config: process_auto_fields", function()
                         config = {
                           http_endpoint = "https://example.com",
                           content_type = "application/json",
-                          flush_timeout = 2,
                           keepalive = 60000,
                           method = "POST",
-                          queue_size = 1,
-                          retry_count = 10,
                           timeout = 10000,
+                          queue = {
+                            initial_retry_delay = 0.01,
+                            max_batch_size = 1,
+                            max_entries = 10000,
+                            max_coalescing_delay = 1,
+                            max_retry_delay = 60,
+                            max_retry_time = 60,
+                          },
                         }
                       }
                     }